feat(docker): share as docker-compose, reset alembic
