+++ .env.example
@@ -1,16 +1,70 @@
-# DB
-MYSQL_DATABASE=your_db_name
-DATABASE_NAME=your_db_user
-DATABASE_PASSWORD=your_db_password
-DATABASE_ROOT_PASSWORD=your_root_password
-DATABASE_URL=mysql+aiomysql://${DATABASE_NAME}:${DATABASE_PASSWORD}@llearn-mysql:3306/${MYSQL_DATABASE}
+# =============================================================================
+# 🚀 LLEARN 플랫폼 환경설정
+# =============================================================================
+# 설치 후 아래 항목들을 실제 값으로 변경해주세요
+# 변경 필요한 항목은 🔥 표시되어 있습니다
+
+# =============================================================================
+# 📊 데이터베이스 설정
+# =============================================================================
+MYSQL_DATABASE=llearn
+MYSQL_USER=llearn_user
+MYSQL_PASSWORD=llearn_password
+
+DATABASE_NAME=llearn_admin
+DATABASE_PASSWORD=llearn_pass123!             # 🔥 보안을 위해 변경하세요
+DATABASE_ROOT_PASSWORD=super_secret_root123!  # 🔥 보안을 위해 변경하세요
+
+# =============================================================================
+# 🏷️ 애플리케이션 설정
+# =============================================================================
+APP_NAME=llearn
+APP_VERSION=1.0.0
+
+# =============================================================================
+# 🔐 보안 설정
+# =============================================================================
+SECRET_KEY=LLEARN_SUPER_SECRET_KEY  # 🔥 32자 이상의 랜덤 문자열로 변경
+ALGORITHM=HS256
+
+# =============================================================================
+# 🤖 AI/LLM 설정
+# =============================================================================
+LLM_API_KEY=your_openai_api_key_here    # 🔥 OpenAI API 키 입력
+LLM_MODEL=gpt-4o                   
+LLM_ENDPOINT=https://api.openai.com/v1/chat/completions
+
+# =============================================================================
+# 👤 관리자 계정 (저장되어있음)
+# =============================================================================
+ADMIN_EMAIL=admin@admin.com
+ADMIN_NAME=시스템관리자
+ADMIN_PASSWORD=!Admin1234
+
+# =============================================================================
+# 🔧 시스템 내부 설정 (일반적으로 변경 불필요)
+# =============================================================================
+# 데이터베이스 연결 URL (자동 생성됨)
+DATABASE_URL=mysql+aiomysql://${DATABASE_NAME}:${DATABASE_PASSWORD}@${MYSQL_DATABASE}-db:3306/${MYSQL_DATABASE}
 SQLALCHEMY_DATABASE_URL=${DATABASE_URL}
 
-# Auth / Crypto
-SECRET_KEY=your_secret_key
-ALGORITHM=HS256
-ADMIN_HASH=your_hashed_password
+# Docker 내부 연결 설정
+DB_HOST=${MYSQL_DATABASE}-db
 
-# LLM (e.g., OpenAI)
-LLM_API_KEY=your_openai_token
-LLM_ENDPOINT=https://api.openai.com/v1/chat/completions
+# =============================================================================
+# 🌐 네트워크 설정
+# =============================================================================
+# 웹 서버 포트
+WEB_PORT=8000
+
+# 모니터링 포트
+PROMETHEUS_PORT=9090
+GRAFANA_PORT=3000
+
+# =============================================================================
+# 📝 설정 가이드
+# =============================================================================
+# 1. 🔥 표시된 항목들을 반드시 변경하세요
+# 2. SECRET_KEY 생성: python -c "import secrets; print(secrets.token_urlsafe(32))"
+# 3. OpenAI API 키: https://platform.openai.com/api-keys
+# 4. 강력한 비밀번호 사용을 권장합니다
