--- alembic/versions/2025_07_23_1414-3211bbd1f621_create_users_table.py
@@ -1,41 +1,0 @@
-"""create users table
-
-Revision ID: 3211bbd1f621
-Revises: 
-Create Date: 2025-07-23 14:14:00.600740
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '3211bbd1f621'
-down_revision: Union[str, Sequence[str], None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('users',
-    sa.Column('id', sa.String(length=26), nullable=False),
-    sa.Column('email', sa.String(length=64), nullable=False),
-    sa.Column('name', sa.String(length=32), nullable=False),
-    sa.Column('password', sa.String(length=64), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=False),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('email')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('users')
-    # ### end Alembic commands ###
