--- alembic/versions/2025_07_24_0024-715448e08026_add_role_column_to_users.py
@@ -1,32 +1,0 @@
-"""Add role column to users
-
-Revision ID: 715448e08026
-Revises: 3211bbd1f621
-Create Date: 2025-07-24 00:24:20.265788
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '715448e08026'
-down_revision: Union[str, Sequence[str], None] = '3211bbd1f621'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('role', sa.Enum('USER', 'ADMIN', name='user_role'), nullable=False))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('users', 'role')
-    # ### end Alembic commands ###
