--- alembic/versions/2025_08_01_0922-529049bbaf14_fix_curriculum_db.py
@@ -1,72 +1,0 @@
-"""fix curriculum db
-
-Revision ID: 529049bbaf14
-Revises: 54e95a0912f4
-Create Date: 2025-08-01 09:22:58.164446
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '529049bbaf14'
-down_revision: Union[str, Sequence[str], None] = '54e95a0912f4'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('curriculums',
-    sa.Column('id', sa.String(length=26), nullable=False),
-    sa.Column('user_id', sa.String(length=26), nullable=False),
-    sa.Column('title', sa.String(length=50), nullable=False),
-    sa.Column('visibility', sa.String(length=10), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('summaries',
-    sa.Column('id', sa.String(length=26), nullable=False),
-    sa.Column('curriculum_id', sa.String(length=26), nullable=False),
-    sa.Column('week_number', sa.Integer(), nullable=False),
-    sa.Column('content', sa.Text(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=False),
-    sa.ForeignKeyConstraint(['curriculum_id'], ['curriculums.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('week_schedules',
-    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('curriculum_id', sa.String(length=26), nullable=False),
-    sa.Column('week_number', sa.Integer(), nullable=False),
-    sa.Column('lessons', sa.JSON(), nullable=False),
-    sa.ForeignKeyConstraint(['curriculum_id'], ['curriculums.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('feedbacks',
-    sa.Column('id', sa.String(length=26), nullable=False),
-    sa.Column('summary_id', sa.String(length=26), nullable=False),
-    sa.Column('comment', sa.Text(), nullable=False),
-    sa.Column('score', sa.Float(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=False),
-    sa.ForeignKeyConstraint(['summary_id'], ['summaries.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('feedbacks')
-    op.drop_table('week_schedules')
-    op.drop_table('summaries')
-    op.drop_table('curriculums')
-    # ### end Alembic commands ###
