--- alembic/versions/2025_08_01_2156-c338b5d92931_add_social_table.py
@@ -1,52 +1,0 @@
-"""add social table
-
-Revision ID: c338b5d92931
-Revises: 529049bbaf14
-Create Date: 2025-08-01 21:56:00.614259
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'c338b5d92931'
-down_revision: Union[str, Sequence[str], None] = '529049bbaf14'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('bookmarks',
-    sa.Column('id', sa.String(length=26), nullable=False),
-    sa.Column('user_id', sa.String(length=26), nullable=False),
-    sa.Column('curriculum_id', sa.String(length=26), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.ForeignKeyConstraint(['curriculum_id'], ['curriculums.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('user_id', 'curriculum_id', name='unique_user_curriculum_bookmark')
-    )
-    op.create_table('likes',
-    sa.Column('id', sa.String(length=26), nullable=False),
-    sa.Column('user_id', sa.String(length=26), nullable=False),
-    sa.Column('curriculum_id', sa.String(length=26), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.ForeignKeyConstraint(['curriculum_id'], ['curriculums.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('user_id', 'curriculum_id', name='unique_user_curriculum_like')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('likes')
-    op.drop_table('bookmarks')
-    # ### end Alembic commands ###
