--- alembic/versions/2025_08_01_2239-9e48abd53979_add_comment_tables.py
@@ -1,53 +1,0 @@
-"""add comment tables
-
-Revision ID: 9e48abd53979
-Revises: c338b5d92931
-Create Date: 2025-08-01 22:39:56.415414
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '9e48abd53979'
-down_revision: Union[str, Sequence[str], None] = 'c338b5d92931'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('comments',
-    sa.Column('id', sa.String(length=26), nullable=False),
-    sa.Column('user_id', sa.String(length=26), nullable=False),
-    sa.Column('curriculum_id', sa.String(length=26), nullable=False),
-    sa.Column('content', sa.Text(), nullable=False),
-    sa.Column('parent_comment_id', sa.String(length=26), nullable=True),
-    sa.Column('is_deleted', sa.Boolean(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=False),
-    sa.ForeignKeyConstraint(['curriculum_id'], ['curriculums.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['parent_comment_id'], ['comments.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index('idx_curriculum_created', 'comments', ['curriculum_id', 'created_at'], unique=False)
-    op.create_index('idx_curriculum_deleted', 'comments', ['curriculum_id', 'is_deleted'], unique=False)
-    op.create_index('idx_parent_created', 'comments', ['parent_comment_id', 'created_at'], unique=False)
-    op.create_index('idx_user_created', 'comments', ['user_id', 'created_at'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index('idx_user_created', table_name='comments')
-    op.drop_index('idx_parent_created', table_name='comments')
-    op.drop_index('idx_curriculum_deleted', table_name='comments')
-    op.drop_index('idx_curriculum_created', table_name='comments')
-    op.drop_table('comments')
-    # ### end Alembic commands ###
