+++ alembic/versions/2025_08_02_1236-c68040e76137_reset_initial_migration.py
@@ -1,0 +1,203 @@
+"""reset & initial migration
+
+Revision ID: c68040e76137
+Revises: 
+Create Date: 2025-08-02 12:36:58.143188
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'c68040e76137'
+down_revision: Union[str, Sequence[str], None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('categories',
+    sa.Column('id', sa.String(length=26), nullable=False),
+    sa.Column('name', sa.String(length=30), nullable=False),
+    sa.Column('description', sa.Text(), nullable=True),
+    sa.Column('color', sa.String(length=7), nullable=False),
+    sa.Column('icon', sa.String(length=50), nullable=True),
+    sa.Column('sort_order', sa.Integer(), nullable=False),
+    sa.Column('is_active', sa.Boolean(), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('name')
+    )
+    op.create_index('idx_category_active_sort', 'categories', ['is_active', 'sort_order'], unique=False)
+    op.create_index('idx_category_name', 'categories', ['name'], unique=False)
+    op.create_table('users',
+    sa.Column('id', sa.String(length=26), nullable=False),
+    sa.Column('email', sa.String(length=64), nullable=False),
+    sa.Column('name', sa.String(length=32), nullable=False),
+    sa.Column('password', sa.String(length=64), nullable=False),
+    sa.Column('role', sa.Enum('USER', 'ADMIN', name='user_role'), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('email')
+    )
+    op.create_table('curriculums',
+    sa.Column('id', sa.String(length=26), nullable=False),
+    sa.Column('user_id', sa.String(length=26), nullable=False),
+    sa.Column('title', sa.String(length=50), nullable=False),
+    sa.Column('visibility', sa.String(length=10), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=False),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('tags',
+    sa.Column('id', sa.String(length=26), nullable=False),
+    sa.Column('name', sa.String(length=20), nullable=False),
+    sa.Column('usage_count', sa.Integer(), nullable=False),
+    sa.Column('created_by', sa.String(length=26), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=False),
+    sa.ForeignKeyConstraint(['created_by'], ['users.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('name')
+    )
+    op.create_index('idx_tag_created_by', 'tags', ['created_by'], unique=False)
+    op.create_index('idx_tag_name', 'tags', ['name'], unique=False)
+    op.create_index('idx_tag_usage_count', 'tags', ['usage_count'], unique=False)
+    op.create_table('bookmarks',
+    sa.Column('id', sa.String(length=26), nullable=False),
+    sa.Column('user_id', sa.String(length=26), nullable=False),
+    sa.Column('curriculum_id', sa.String(length=26), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.ForeignKeyConstraint(['curriculum_id'], ['curriculums.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('user_id', 'curriculum_id', name='unique_user_curriculum_bookmark')
+    )
+    op.create_table('comments',
+    sa.Column('id', sa.String(length=26), nullable=False),
+    sa.Column('user_id', sa.String(length=26), nullable=False),
+    sa.Column('curriculum_id', sa.String(length=26), nullable=False),
+    sa.Column('content', sa.Text(), nullable=False),
+    sa.Column('parent_comment_id', sa.String(length=26), nullable=True),
+    sa.Column('is_deleted', sa.Boolean(), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=False),
+    sa.ForeignKeyConstraint(['curriculum_id'], ['curriculums.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['parent_comment_id'], ['comments.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index('idx_curriculum_created', 'comments', ['curriculum_id', 'created_at'], unique=False)
+    op.create_index('idx_curriculum_deleted', 'comments', ['curriculum_id', 'is_deleted'], unique=False)
+    op.create_index('idx_parent_created', 'comments', ['parent_comment_id', 'created_at'], unique=False)
+    op.create_index('idx_user_created', 'comments', ['user_id', 'created_at'], unique=False)
+    op.create_table('curriculum_categories',
+    sa.Column('id', sa.String(length=26), nullable=False),
+    sa.Column('curriculum_id', sa.String(length=26), nullable=False),
+    sa.Column('category_id', sa.String(length=26), nullable=False),
+    sa.Column('assigned_by', sa.String(length=26), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.ForeignKeyConstraint(['assigned_by'], ['users.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['category_id'], ['categories.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['curriculum_id'], ['curriculums.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('curriculum_id', name='unique_curriculum_category')
+    )
+    op.create_index('idx_curriculum_category_assigned_by', 'curriculum_categories', ['assigned_by'], unique=False)
+    op.create_index('idx_curriculum_category_category', 'curriculum_categories', ['category_id'], unique=False)
+    op.create_index('idx_curriculum_category_curriculum', 'curriculum_categories', ['curriculum_id'], unique=False)
+    op.create_table('curriculum_tags',
+    sa.Column('id', sa.String(length=26), nullable=False),
+    sa.Column('curriculum_id', sa.String(length=26), nullable=False),
+    sa.Column('tag_id', sa.String(length=26), nullable=False),
+    sa.Column('added_by', sa.String(length=26), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.ForeignKeyConstraint(['added_by'], ['users.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['curriculum_id'], ['curriculums.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['tag_id'], ['tags.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('curriculum_id', 'tag_id', name='unique_curriculum_tag')
+    )
+    op.create_index('idx_curriculum_tag_added_by', 'curriculum_tags', ['added_by'], unique=False)
+    op.create_index('idx_curriculum_tag_curriculum', 'curriculum_tags', ['curriculum_id'], unique=False)
+    op.create_index('idx_curriculum_tag_tag', 'curriculum_tags', ['tag_id'], unique=False)
+    op.create_table('likes',
+    sa.Column('id', sa.String(length=26), nullable=False),
+    sa.Column('user_id', sa.String(length=26), nullable=False),
+    sa.Column('curriculum_id', sa.String(length=26), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.ForeignKeyConstraint(['curriculum_id'], ['curriculums.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('user_id', 'curriculum_id', name='unique_user_curriculum_like')
+    )
+    op.create_table('summaries',
+    sa.Column('id', sa.String(length=26), nullable=False),
+    sa.Column('curriculum_id', sa.String(length=26), nullable=False),
+    sa.Column('week_number', sa.Integer(), nullable=False),
+    sa.Column('content', sa.Text(), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=False),
+    sa.ForeignKeyConstraint(['curriculum_id'], ['curriculums.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('week_schedules',
+    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('curriculum_id', sa.String(length=26), nullable=False),
+    sa.Column('week_number', sa.Integer(), nullable=False),
+    sa.Column('lessons', sa.JSON(), nullable=False),
+    sa.ForeignKeyConstraint(['curriculum_id'], ['curriculums.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('feedbacks',
+    sa.Column('id', sa.String(length=26), nullable=False),
+    sa.Column('summary_id', sa.String(length=26), nullable=False),
+    sa.Column('comment', sa.Text(), nullable=False),
+    sa.Column('score', sa.Float(), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=False),
+    sa.ForeignKeyConstraint(['summary_id'], ['summaries.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('feedbacks')
+    op.drop_table('week_schedules')
+    op.drop_table('summaries')
+    op.drop_table('likes')
+    op.drop_index('idx_curriculum_tag_tag', table_name='curriculum_tags')
+    op.drop_index('idx_curriculum_tag_curriculum', table_name='curriculum_tags')
+    op.drop_index('idx_curriculum_tag_added_by', table_name='curriculum_tags')
+    op.drop_table('curriculum_tags')
+    op.drop_index('idx_curriculum_category_curriculum', table_name='curriculum_categories')
+    op.drop_index('idx_curriculum_category_category', table_name='curriculum_categories')
+    op.drop_index('idx_curriculum_category_assigned_by', table_name='curriculum_categories')
+    op.drop_table('curriculum_categories')
+    op.drop_index('idx_user_created', table_name='comments')
+    op.drop_index('idx_parent_created', table_name='comments')
+    op.drop_index('idx_curriculum_deleted', table_name='comments')
+    op.drop_index('idx_curriculum_created', table_name='comments')
+    op.drop_table('comments')
+    op.drop_table('bookmarks')
+    op.drop_index('idx_tag_usage_count', table_name='tags')
+    op.drop_index('idx_tag_name', table_name='tags')
+    op.drop_index('idx_tag_created_by', table_name='tags')
+    op.drop_table('tags')
+    op.drop_table('curriculums')
+    op.drop_table('users')
+    op.drop_index('idx_category_name', table_name='categories')
+    op.drop_index('idx_category_active_sort', table_name='categories')
+    op.drop_table('categories')
+    # ### end Alembic commands ###
