+++ docker-compose.yml
@@ -3,7 +3,7 @@ version: "3.9"
 services:
   db:
     image: mysql:8
-    container_name: llearn-mysql
+    container_name: "${MYSQL_DATABASE}-db"
     restart: unless-stopped
     environment:
       MYSQL_DATABASE: "${MYSQL_DATABASE}"
@@ -23,14 +23,13 @@ services:
   app:
     build:
       context: .
-      dockerfile: Dockerfile        # 아직 비어 있어도 OK – 추후 작성
-    container_name: llearn-app
+      dockerfile: Dockerfile
+    container_name: "${APP_NAME}-app"
     restart: unless-stopped
     env_file: .env
     depends_on:
       db:
         condition: service_healthy
-    command: uvicorn main:app --host 0.0.0.0 --port 8000 --reload
     volumes:
       - .:/workspace
     ports:
@@ -39,7 +38,7 @@ services:
 
   prometheus:
     image: prom/prometheus:v2.45.0
-    container_name: llearn-prometheus
+    container_name: "${APP_NAME}-prometheus"
     restart: unless-stopped
     ports:
       - "9090:9090"
@@ -58,7 +57,7 @@ services:
 
   grafana:
     image: grafana/grafana:10.0.0
-    container_name: llearn-grafana
+    container_name: "${APP_NAME}-grafana"
     restart: unless-stopped
     ports:
       - "3000:3000"
